requirements.txt
streamlit 
openai 
pyyaml
python-dotenv
fastapi
uvicorn
pydantic
pyyaml
requests